rust: use i32 type in Value

+++ bindings/rust/src/lib.rs
@@ -79,10 +79,10 @@ pub type Value = sys::FizzyValue;
 // NOTE: the union does not have i32
 impl Value {
     pub fn as_i32(&self) -> i32 {
-        unsafe { self.i64 as i32 }
+        unsafe { self.i32 as i32 }
     }
     pub fn as_u32(&self) -> u32 {
-        unsafe { self.i64 as u32 }
+        unsafe { self.i32 }
     }
     pub fn as_i64(&self) -> i64 {
         unsafe { self.i64 as i64 }
@@ -100,13 +100,13 @@ impl Value {
 
 impl From<i32> for Value {
     fn from(v: i32) -> Self {
-        Value { i64: v as u64 }
+        Value { i32: v as u32 }
     }
 }
 
 impl From<u32> for Value {
     fn from(v: u32) -> Self {
-        Value { i64: v as u64 }
+        Value { i32: v }
     }
 }
